Enhance tests for NNMC catalog synchronization and program repository
- Updated test_sync_nnmc_catalog.py to reflect changes in the catalog, including:
  - Adjusted expected course count and added assertions for course uniqueness.
  - Enhanced program_identity tests to match official catalog names and credentials.
  - Added tests for program credit range and course code/title matching.
  - Implemented validation for duplicate official IDs in the catalog snapshot.
  - Created fake database and connection classes for testing database interactions.
  - Added idempotency test for catalog synchronization with rollback on failure.

- Modified test_program_repository.py to include additional fields in program summaries:
  - Added total_credit_hours_text, total_credit_hours_min, total_credit_hours_max,
    official_source_id, official_source_url, and source_retrieved_at.

+++ fastapi_backend/app/student/domains/planning/schemas/education.py
@@ -17,7 +17,7 @@ class ProgramCoursePayload(BaseModel):
     semester: str | None = None
     code: str | None = None
     course_name: str | None = Field(None, alias="courseName")
-    credits: int | None = None
+    credits: float | None = None
     prerequisite: str | None = None
     corequisite: str | None = None
     schedule: ScheduleBlock | dict | None = None
